pyproject.toml
[project]
name = "mcp-weather"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.24.0",
    "python-dotenv>=1.2.1",
]
